Options, Blacklisting, Comments, Reputation, Various fixes
Added account options and default values for when not logged in, as well as giving the option to set the default rating during the installation.

Added the ability to blacklist tags on the main browse page.

Added the ability to add reputation for a user with a note.

Added the ability to comment and edit posted comment on a post. As well as voting a post. Reporting it and managing it in the admin area as well.