Comment page, Various fixes
Added browsing comments in extra.php?a=c finally! You can also search for users with user:{$user}

Fixed the comment and deleted comment count on profile page.

FIxed the thumbnail of videos having too high resolutions, so I resize them after they have been generated.

Fixed a bug with the blacklist. If a user hasn't edited it once, it would be null in the db and result in an unreadable string in posts.php line 55, added ?? "" to ensure nothing happens in that scenario.

Fixed blacklist issue #11 by simply moving all the tags with a minus at the end of the tags array.